% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{safeSQLVar}
\alias{safeSQLVar}
\title{Safe SQL String}
\usage{
safeSQLVar(var)
}
\arguments{
\item{var}{String to be made safe}
}
\value{
Safe sql string
}
\description{
Safe SQL String
}
\examples{
safeSQLVar("unsafe' string")
}
